Separate config namespaces
